Renamed methods in replacements; created redis package in foo/somedifferentpackage/examples/methodreplacement/; renamed addRedisInfo method to addRedisCommand
