+++ client-java/instrumentation/src/main/java/org/evomaster/client/java/instrumentation/AdditionalInfo.java
@@ -150,7 +150,7 @@ public class AdditionalInfo implements Serializable {
         openSearchCommandData.add(info);
     }
 
-    public void addRedisInfo(RedisCommand info){
+    public void addRedisCommand(RedisCommand info){
         redisCommandData.add(info);
     }
