+++ client-java/instrumentation/src/main/java/org/evomaster/client/java/instrumentation/staticstate/ExecutionTracer.java
@@ -433,9 +433,9 @@ public class ExecutionTracer {
         getCurrentAdditionalInfo().addOpenSearchInfo(info);
     }
 
-    public static void addRedisInfo(RedisCommand info){
+    public static void addRedisCommand(RedisCommand info){
         if (!executingInitRedis)
-            getCurrentAdditionalInfo().addRedisInfo(info);
+            getCurrentAdditionalInfo().addRedisCommand(info);
     }
 
     public static void addMongoCollectionType(MongoCollectionSchema mongoCollectionSchema){
