+++ client-java/instrumentation/src/test/java/com/foo/somedifferentpackage/examples/methodreplacement/redis/RedisValueOperationsImpl.java
@@ -1,7 +1,6 @@
-package com.foo.somedifferentpackage.examples.methodreplacement;
+package com.foo.somedifferentpackage.examples.methodreplacement.redis;
 
 import org.evomaster.client.java.instrumentation.example.redis.RedisValueOperations;
-import org.springframework.data.redis.core.RedisTemplate;
 
 import java.lang.reflect.Method;
